1384 Removes unused helper functions
